+++ RELEASES.md
@@ -1,5 +1,6 @@
 Version 0.8.7 (2021-XX-XX)
 ========================
+ * Volkswagen T-Cross 2021 support thanks to jyoung8607!
 
 Version 0.8.6 (2021-07-21)
 ========================